cdk/lambda/practiceMaterial/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via langchain-community
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via httpx
asyncpg==0.31.0
    # via langchain-postgres
attrs==25.4.0
    # via aiohttp
boto3==1.42.31
    # via
    #   -r requirements.in
    #   langchain-aws
botocore==1.42.31
    # via
    #   -r requirements.in
    #   boto3
    #   s3transfer
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
dataclasses-json==0.6.7
    # via langchain-community
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.3.0
    # via sqlalchemy
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   langgraph-sdk
    #   langsmith
httpx-sse==0.4.3
    # via langchain-community
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==1.2.6
    # via -r requirements.in
langchain-aws==1.0.0
    # via -r requirements.in
langchain-classic==1.0.1
    # via
    #   -r requirements.in
    #   langchain-community
langchain-community==0.4.1
    # via -r requirements.in
langchain-core==1.2.7
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-aws
    #   langchain-classic
    #   langchain-community
    #   langchain-postgres
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-postgres==0.0.16
    # via -r requirements.in
langchain-text-splitters==1.1.0
    # via langchain-classic
langgraph==1.0.6
    # via langchain
langgraph-checkpoint==4.0.0
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.6
    # via langgraph
langgraph-sdk==0.3.3
    # via langgraph
langsmith==0.6.4
    # via
    #   langchain-classic
    #   langchain-community
    #   langchain-core
marshmallow==3.26.2
    # via dataclasses-json
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
numpy==2.4.1
    # via
    #   -r requirements.in
    #   langchain-aws
    #   langchain-community
    #   langchain-postgres
    #   pgvector
orjson==3.11.5
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.12.2
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   langchain-core
    #   langsmith
    #   marshmallow
pgvector==0.3.6
    # via langchain-postgres
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
psycopg==3.3.2
    # via
    #   -r requirements.in
    #   langchain-postgres
psycopg-binary==3.3.2
    # via psycopg
psycopg-pool==3.3.0
    # via
    #   langchain-postgres
    #   psycopg
psycopg2-binary==2.9.11
    # via -r requirements.in
pydantic==2.12.5
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-aws
    #   langchain-classic
    #   langchain-core
    #   langgraph
    #   langsmith
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via langchain-community
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.2.1
    # via pydantic-settings
pyyaml==6.0.3
    # via
    #   langchain-classic
    #   langchain-community
    #   langchain-core
requests==2.32.5
    # via
    #   langchain-classic
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
s3transfer==0.16.0
    # via boto3
six==1.17.0
    # via python-dateutil
sqlalchemy==2.0.46
    # via
    #   langchain-classic
    #   langchain-community
    #   langchain-postgres
tenacity==9.1.2
    # via
    #   langchain-community
    #   langchain-core
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   langchain-core
    #   psycopg
    #   psycopg-pool
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.3
    # via psycopg
urllib3==2.6.3
    # via
    #   -r requirements.in
    #   botocore
    #   requests
uuid-utils==0.14.0
    # via
    #   langchain-core
    #   langsmith
xxhash==3.6.0
    # via langgraph
yarl==1.22.0
    # via aiohttp
zstandard==0.25.0
    # via langsmith
